+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## master (unreleased)
+* Show discount fields straight from the start, not after entering date
+* Make Discounts section a `form` and add a wrapper to be able to align label and input without affecting the button and message
+* Update default CSS
+
 ## 0.18.0 (2019-09-28)
 * Add heading to quantity form
 * Add placeholders to quantity form inputs